refactor: Move to MenuRenderer
