+++ lib/components/OverflowMenu/OverflowMenu.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { menuTestSuite } from '../Menu/testHelper';
+import { menuTestSuite } from '../MenuRenderer/testHelper';
 import { OverflowMenu } from './OverflowMenu';
 
 menuTestSuite({
